refactor(core/domain): promote PositionTranslator interface from jvmMain to commonMain
